Fix build with clang 3.4
This fixes various build failures with clang 3.4, such as hardcoded
assumptions about clang based on clang 3.3, making sure we don't pass
unnamed types to templates, and making sure snprintf is #undef-ed before
being touched by string_util.h.

+++ content/browser/android/touch_point.cc
@@ -32,7 +32,7 @@ void MaybeAddTouchPoint(JNIEnv* env,
 
   // Record the current number of points in the WebTouchEvent
   const int idx = event.touchesLength;
-  DCHECK_LT(idx, WebKit::WebTouchEvent::touchesLengthCap);
+  DCHECK_LT(idx, (int)WebKit::WebTouchEvent::touchesLengthCap);
 
   WebTouchPoint wtp;
   wtp.id = Java_TouchPoint_getId(env, pt);
